# Switch count type on map with no arguments to BPF_MAP_TYPE_PERCPU_ARRAY